+++ examples/server/vite.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'vite'
 import plugin from 'phecda-client/vite'
 
 export default defineConfig({
-  plugins: [plugin()],
+  plugins: [plugin({ split: true })],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url)),