+++ packages/server/__test__/core.spec.ts
@@ -17,6 +17,27 @@ describe('Factory ', () => {
     expect(data).toMatchSnapshot()
   })
 
+  it('In extended case', async () => {
+    @Controller('/A')// It won't work
+    class A {
+      @Get('/testA')
+      test(@Query('id') id: string, @Body('name') name: string) {
+        return id + name
+      }
+    }
+
+    @Controller('/B')
+    class B extends A {
+      @Get('/testB')
+      test2(@Query('id') id: string, @Body('name') name: string) {
+        return id + name
+      }
+    }
+    const { meta } = await Factory([B])
+    const data = meta.map(item => item.data)
+    expect(data).toMatchSnapshot()
+  })
+
   it('Factory will work using nest class', async () => {
     class Service {
       test() {